// frontend/src/types/generated/models/TestCase.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { DataSplit } from "./DataSplit";
import type { JsonValue } from "../serde_json/JsonValue";
import type { TaskReference } from "./TaskReference";

/**
 * 测试用例结构
 */
export type TestCase = { 
/**
 * 唯一标识
 */
id: string, 
/**
 * 输入变量
 */
input: { [key in string]?: JsonValue }, 
/**
 * 期望输出/约束
 */
reference: TaskReference, 
/**
 * 数据划分归属
 */
split: DataSplit | null, 
/**
 * 元数据
 */
metadata: { [key in string]?: JsonValue } | null, };